requirements.txt
alembic==1.7.5
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
undetected-chromedriver
certifi==2022.6.15
opencv-python
pillow
cffi==1.15.0
charset-normalizer==2.0.10
click==7.1.2
et-xmlfile==1.1.0
fake-useragent==0.1.11
Flask==1.1.4
Flask-Bcrypt==0.7.1
Flask-Migrate==2.1.1
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.4.0
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.1
openpyxl==3.0.9
pandas==1.3.5
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
urllib3==1.26.8
Werkzeug==0.16.1
zipp==3.7.0
